Update readme, docs, improve benchmarker
The benchmarker now has a better system (a hashmap with named timers)
to profile different parts of the frame, and it has already made one
fact clear: on Windows, Glutin builds seem to do the "vsync hang" in
the renderer, instead of during swap_buffers. Which is very odd. Time
to investigate.