Switch to a firefox comparison and reverse srgb
Turns out that the whole "the gimp rendered text looks like srgb" was
caused by the fact that the gimp rendered image was srgb, so of course
it looked like srgb. Basically, don't draw text with gimp without paying
careful attention to your color spaces.

Firefox will be used as a comparison target in `text_layout` from now
on, as it's the premier text reading application today anyways.